Refactor social links management: move social links loading logic to SocialLinksFooter component and remove redundant code from App component
